rename test to l1_observer_test
